DECK ORDERED BY VALUE
Deck objects created with card array in ascending order of card value.

This change was made by redefining the correspondence between card_id and rank_id and card_id and suit_id as defined in card class. Now card value is the same as card id. This also means the value no longer needs to be calculated as it is just the card_id.

An additional fix was applied to deck.removecards method to work correctly on an array of card to remove.

New methods and attributes tested indivdually and in games.ipynb